ClassTestCase.
- Fixed naming of several tests(eg dropped Class in test name).
- CharSequenceTestCase base class updated, removed extra ClassTests.
- SpreadsheetParserTestCase base class updated ParserTestCase -> ParserTestCase3.
- Updated tests to use expected= rather than assert exception thrown.
