+++ src/main/java/walkingkooka/text/PaddedCharSequence.java
