+++ src/test/java/walkingkooka/compare/ComparisonRelationTest.java
@@ -18,8 +18,9 @@
 
 package walkingkooka.compare;
 
+import org.junit.Ignore;
 import org.junit.Test;
-import walkingkooka.test.PublicClassTestCase;
+import walkingkooka.predicate.PredicateTestCase;
 
 import java.util.function.Predicate;
 
@@ -30,7 +31,202 @@ import static org.junit.Assert.assertNotSame;
 import static org.junit.Assert.assertSame;
 import static org.junit.Assert.assertTrue;
 
-public final class ComparisonRelationTest extends PublicClassTestCase<ComparisonRelation> {
+public final class ComparisonRelationTest extends PredicateTestCase<ComparisonRelation, Integer> {
+
+    @Test
+    @Ignore
+    @Override
+    public void testCheckNaming() {
+        throw new UnsupportedOperationException();
+    }
+
+    @Test
+    @Ignore
+    @Override
+    public void testClassVisibility() {
+        throw new UnsupportedOperationException();
+    }
+
+    @Test
+    @Ignore
+    @Override
+    public void testAllMethodsVisibility() {
+        throw new UnsupportedOperationException();
+    }
+
+    // EQ................................................................
+
+    @Test
+    public void testEqNegative() {
+        this.testFalse2(ComparisonRelation.EQ, Integer.MIN_VALUE);
+    }
+
+    @Test
+    public void testEqNegative2() {
+        this.testFalse2(ComparisonRelation.EQ,-1);
+    }
+
+    @Test
+    public void testEqZero() {
+        this.testTrue2(ComparisonRelation.EQ,0);
+    }
+
+    @Test
+    public void testEqPositive() {
+        this.testFalse2(ComparisonRelation.EQ,+1);
+    }
+
+    @Test
+    public void testEqPositive2() {
+        this.testFalse2(ComparisonRelation.EQ, Integer.MAX_VALUE);
+    }
+
+    // GT................................................................
+
+    @Test
+    public void testGtNegative() {
+        this.testFalse2(ComparisonRelation.GT, Integer.MIN_VALUE);
+    }
+
+    @Test
+    public void testGtNegative2() {
+        this.testFalse2(ComparisonRelation.GT, -1);
+    }
+
+    @Test
+    public void testGtZero() {
+        this.testFalse2(ComparisonRelation.GT, 0);
+    }
+
+    @Test
+    public void testGtPositive() {
+        this.testTrue2(ComparisonRelation.GT, +1);
+    }
+
+    @Test
+    public void testGtPositive2() {
+        this.testTrue2(ComparisonRelation.GT, Integer.MAX_VALUE);
+    }
+
+    // GTE................................................................
+
+    @Test
+    public void testGteNegative() {
+        this.testFalse2(ComparisonRelation.GTE, Integer.MIN_VALUE);
+    }
+
+    @Test
+    public void testGteNegative2() {
+        this.testFalse2(ComparisonRelation.GTE, -1);
+    }
+
+    @Test
+    public void testGteZero() {
+        this.testTrue2(ComparisonRelation.GTE, 0);
+    }
+
+    @Test
+    public void testGtePositive() {
+        this.testTrue2(ComparisonRelation.GTE, +1);
+    }
+
+    @Test
+    public void testGtePositive2() {
+        this.testTrue2(ComparisonRelation.GTE, Integer.MAX_VALUE);
+    }
+
+    // LT................................................................
+
+    @Test
+    public void testLtNegative() {
+        this.testTrue2(ComparisonRelation.LT, Integer.MIN_VALUE);
+    }
+
+    @Test
+    public void testLtNegative2() {
+        this.testTrue2(ComparisonRelation.LT, -1);
+    }
+
+    @Test
+    public void testLtZero() {
+        this.testFalse2(ComparisonRelation.LT, 0);
+    }
+
+    @Test
+    public void testLtPositive() {
+        this.testFalse2(ComparisonRelation.LT, +1);
+    }
+
+    @Test
+    public void testLtPositive2() {
+        this.testFalse2(ComparisonRelation.LT, Integer.MAX_VALUE);
+    }
+
+    // LTE................................................................
+
+    @Test
+    public void testLteNegative() {
+        this.testTrue2(ComparisonRelation.LTE, Integer.MIN_VALUE);
+    }
+
+    @Test
+    public void testLteNegative2() {
+        this.testTrue2(ComparisonRelation.LTE, -1);
+    }
+
+    @Test
+    public void testLteZero() {
+        this.testTrue2(ComparisonRelation.LTE, 0);
+    }
+
+    @Test
+    public void testLtePositive() {
+        this.testFalse2(ComparisonRelation.LTE, +1);
+    }
+
+    @Test
+    public void testLtePositive2() {
+        this.testFalse2(ComparisonRelation.LTE, Integer.MAX_VALUE);
+    }
+
+    // NE................................................................
+
+    @Test
+    public void testNeNegative() {
+        this.testTrue2(ComparisonRelation.NE, Integer.MIN_VALUE);
+    }
+
+    @Test
+    public void testNeNegative2() {
+        this.testTrue2(ComparisonRelation.NE, -1);
+    }
+
+    @Test
+    public void testNeZero() {
+        this.testFalse2(ComparisonRelation.NE, 0);
+    }
+
+    @Test
+    public void testNePositive() {
+        this.testTrue2(ComparisonRelation.NE, +1);
+    }
+
+    @Test
+    public void testNePositive2() {
+        this.testTrue2(ComparisonRelation.NE, Integer.MAX_VALUE);
+    }
+
+    // helpers............................................................
+
+    private void testTrue2(final ComparisonRelation relation, final int value) {
+        this.testTrue(relation, value);
+        this.testFalse(relation.invert(), value);
+    }
+
+    private void testFalse2(final ComparisonRelation relation, final int value) {
+        this.testFalse(relation, value);
+        this.testTrue(relation.invert(), value);
+    }
 
     @Test
     public void testSymbolEQ() {
@@ -142,6 +338,11 @@ public final class ComparisonRelationTest extends PublicClassTestCase<Comparison
     @Test
     public void testToString() {
         assertEquals("EQ", ComparisonRelation.EQ.toString());
+    }
+
+    @Override
+    protected ComparisonRelation createPredicate() {
+        return ComparisonRelation.EQ;
     }
 
     @Override
