+++ src/test/java/walkingkooka/text/CharSequenceSubSequencerTest.java
@@ -17,11 +17,50 @@
 
 package walkingkooka.text;
 
+import org.junit.Assert;
+import org.junit.Test;
 import walkingkooka.test.PackagePrivateStaticHelperTestCase;
+
+import static org.junit.Assert.assertEquals;
 
 final public class CharSequenceSubSequencerTest
         extends PackagePrivateStaticHelperTestCase<CharSequenceSubSequencer> {
 
+    @Test(expected = NullPointerException.class)
+    public void testNullFails() {
+        CharSequenceSubSequencer.make(null, 1, 1);
+    }
+
+    @Test
+    public void testToBeforeFromFails() {
+        final int from = 4;
+        final int to = -5;
+        final String string = "string";
+        try {
+            CharSequenceSubSequencer.make(string, from, to);
+            Assert.fail();
+        } catch (final IndexOutOfBoundsException expected) {
+            assertEquals("message",
+                    CharSequenceSubSequencer.toIndexBeforeFromIndex(from, to, string.length()),
+                    expected.getMessage());
+        }
+    }
+
+    @Test
+    public void testMake() {
+        assertEquals("sub", CharSequenceSubSequencer.make(" sub ", 1, 4));
+    }
+
+    @Test
+    public void testNegativeTo() {
+        assertEquals("bcd", CharSequenceSubSequencer.make("abcdef", 1, -2));
+    }
+
+    @Test
+    public void testZeroTo() {
+        assertEquals("bcdef", CharSequenceSubSequencer.make("abcdef", 1, 0));
+    }
+    
     @Override
     protected Class<CharSequenceSubSequencer> type() {
         return CharSequenceSubSequencer.class;
